requirements.txt.txt
langchain
langchain-community
langchain-huggingface
faiss-cpu
huggingface_hub
sentence-transformers
python-dotenv
speechrecognition
pyttsx3
pyaudio